+++ lib/mongoid/persistable.rb
@@ -6,6 +6,7 @@ require "mongoid/persistable/destroyable"
 require "mongoid/persistable/incrementable"
 require "mongoid/persistable/logical"
 require "mongoid/persistable/poppable"
+require "mongoid/persistable/pullable"
 require "mongoid/persistable/renamable"
 require "mongoid/persistable/savable"
 require "mongoid/persistable/settable"
@@ -28,6 +29,7 @@ module Mongoid
     include Incrementable
     include Logical
     include Poppable
+    include Pullable
     include Renamable
     include Savable
     include Settable